+++ workers/tests/unittests/test_report_building_report.py
@@ -12,13 +12,12 @@ class TestReportBuildingReport(unittest.TestCase):
 
         geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
         report = BuildingReport(geometry)
-        report.set_indicator_layer()
 
         # Mock indicator objects with a fixed result value
         for _ in report.indicator_layer:
             indicator = Mock()
             indicator.result = Mock()
-            indicator.result.value = 0.5
+            indicator.result.class_ = 1
             indicator.result.html = "foo"
             report.indicators.append(indicator)
 
@@ -28,7 +27,7 @@ class TestReportBuildingReport(unittest.TestCase):
         self.assertIsNotNone(report.result.label)
         self.assertIsNotNone(report.result.description)
         self.assertIsNotNone(report.result.html)
-        self.assertEqual(report.result.value, 0.5)
+        self.assertEqual(report.result.class_, 1)
 
 
 if __name__ == "__main__":