+++ workers/tests/unittests/test_report_map_action_poc.py
@@ -10,14 +10,13 @@ class TestReportMapActionPoc(unittest.TestCase):
     def test_combine_indicators_mean(self):
         geometry = get_geojson_fixture("heidelberg-altstadt-geometry.geojson")
         report = MapActionPoc(geometry)
-        report.set_indicator_layer()
 
         # Mock indicator objects with a fixed result value
         for _ in report.indicator_layer:
             indicator = Mock()
             indicator.result = Mock()
-            indicator.result.value = 0.5
             indicator.result.html = "foo"
+            indicator.result.class_ = 1
             report.indicators.append(indicator)
 
         report.combine_indicators()
@@ -26,7 +25,7 @@ class TestReportMapActionPoc(unittest.TestCase):
         self.assertIsNotNone(report.result.label)
         self.assertIsNotNone(report.result.description)
         self.assertIsNotNone(report.result.html)
-        self.assertEqual(report.result.value, 0.5)
+        self.assertEqual(report.result.class_, 1)
 
 
 if __name__ == "__main__":